+++ components/event-preview/event-preview-skeleton.vue
