+++ components/event-preview/event-preview.vue
@@ -58,7 +58,7 @@
 </template>
 
 <script>
-import TextDescription from '@/components/eventPreview/textDescription'
+import TextDescription from '@/components/event-preview/text-description'
 
 export default {
   components: {
