+++ components/event-preview/text-description.vue
