+++ components/event/event-skeleton.vue
