+++ pages/events/_eventId.vue
@@ -33,8 +33,8 @@
 <script>
 import Carousel from '@/components/carousel/carousel'
 import CarouselSlide from '@/components/carousel/carousel-slide'
-import EventSkeleton from '@/components/event/eventSkeleton'
-import SpeakerCard from '@/components/event/speakerCard'
+import EventSkeleton from '@/components/event/event-skeleton'
+import SpeakerCard from '@/components/event/speaker-card'
 import Sponsors from '@/components/event/sponsors'
 
 export default {
