+++ pages/index.vue
@@ -32,8 +32,8 @@
 <script>
 /* eslint-disable nuxt/no-this-in-fetch-data */
 
-import eventPreview from '@/components/eventPreview/eventPreview'
-import eventPreviewSkeleton from '@/components/eventPreview/eventPreviewSkeleton'
+import eventPreview from '@/components/event-preview/event-preview'
+import eventPreviewSkeleton from '@/components/event-preview/event-preview-skeleton'
 export default {
   name: 'HomePage',
   components: {
