+++ pages/search.vue
@@ -44,8 +44,8 @@
 /* eslint-disable no-console */
 /* eslint-disable camelcase */
 
-import eventPreviewSkeleton from '@/components/eventPreview/eventPreviewSkeleton'
-import eventPreview from '@/components/eventPreview/eventPreview'
+import eventPreviewSkeleton from '@/components/event-preview/event-preview-skeleton'
+import eventPreview from '@/components/event-preview/event-preview'
 
 export default {
   components: {
